Change datatype for OpeningBalance and OpeningRate

+++ MyPersonalAccountsModel/Inventory/StockItem.cs
@@ -11,7 +11,7 @@ namespace com.techphernalia.MyPersonalAccounts.Model.Inventory
         public string StockItemName { get; set; }
         public int StockGroupId { get; set; }
         public int StockUnitId { get; set; }
-        public decimal OpeningBalance { get; set; }
-        public decimal OpeningRate { get; set; }
+        public double OpeningBalance { get; set; }
+        public double OpeningRate { get; set; }
     }
 }
